Implemented undo and ground transforms
Implemented an undo/redo stack navigated by the z and y keys. It stores the figure, all
views, and the focus iterator for each change. The current undo level and the total number
of levels are displayed in the status area. It functions, but it could be more elegant.

Holding the shift key applies transformations to all figures. Currently flipping and
rotating is applied to each figure in place. I'd like to transform them about the grid
center.